+++ src-electron/db/query-package.js
@@ -187,6 +187,38 @@ async function getPackageByPackageId(db, packageId) {
 }
 
 /**
+ * Returns a package ref from the attribute ID
+ *
+ * @param db
+ * @param attributeId
+ * @returns package ref
+ */
+async function getPackageRefByAttributeId(db, attributeId) {
+  let package_ref = await dbApi.dbAll(
+    db,
+    'SELECT PACKAGE_REF FROM ATTRIBUTE WHERE ATTRIBUTE_ID = ?',
+    [attributeId]
+  )
+  return package_ref[0].PACKAGE_REF
+}
+
+/**
+ * Returns path to the meta file from the package Id
+ *
+ * @param db
+ * @param packageId
+ * @returns Meta file path
+ */
+async function getMetaFile(db, packageId) {
+  let path = await dbApi.dbAll(
+    db,
+    'SELECT PATH FROM PACKAGE WHERE PACKAGE_ID = ?',
+    [packageId]
+  )
+  return path[0].PATH
+}
+
+/**
  * Resolves with a CRC or null for a given path.
  *
  * @export
@@ -979,6 +1011,8 @@ async function insertSessionKeyValuesFromPackageDefaults(db, sessionId) {
 }
 
 // exports
+exports.getMetaFile = getMetaFile
+exports.getPackageRefByAttributeId = getPackageRefByAttributeId
 exports.getPackageByPathAndParent = getPackageByPathAndParent
 exports.getPackageByPackageId = getPackageByPackageId
 exports.getPackagesByType = getPackagesByType
